auth/requirements.txt
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.3
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.1
setuptools==69.1.0
